Move esp.web.util.globaltags to esp.utils.globaltags
I'd prefer to get rid of it, but that's hard.
