+++ esp/esp/context_processors.py
@@ -1,4 +1,4 @@
-import esp.web.util.globaltags
+import esp.utils.globaltags
 from django.contrib.sites.models import Site
 from django.conf import settings
